Refactor JavaScript-Python bridge for Pyodide integration
- Extract Python code into separate module files in app/pycode/
- Simplify JS bridge code by importing Python as raw strings
- Update pyodide.util.js to use modular Python code
- Add .dockerignore file for container builds
- Add pyodide-whl.zip to .gitignore
