+++ .gitignore
@@ -30,6 +30,8 @@ heat-tmp
 
 /heat-stack/other/cache.db
 
+# pyodide files
+/heat-stack/public/pyodide-env/pyodide-whl.zip
 
 # generated files
 /heat-stack/app/components/ui/icons
