+++ heat-stack/.dockerignore
@@ -1,0 +1,25 @@
+# flyctl launch added from .gitignore
+**/node_modules
+**/.DS_store
+
+build
+public/build
+server-build
+**/.env
+
+prisma/data.db
+prisma/data.db-journal
+tests/prisma
+
+test-results
+playwright-report
+playwright/.cache
+tests/fixtures/email
+coverage
+
+other/cache.db
+
+# Easy way to create temporary files/folders that won't accidentally be added to git
+**/*.local.*
+**/public/pyodide-env/rules_engine-*-py3-none-any.whl
+fly.toml
